pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "evo-agent"
version = "0.1.0"
description = "Self-evolving AI agent system"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["extensions*", "core*"]
